+++ Proyecto/asp/ForoUdec mayo 2018/Persistence/correo_institucional.cs
@@ -15,7 +15,7 @@ namespace Persistence
     public partial class correo_institucional
     {
         public int id { get; set; }
-        public Nullable<long> id_session { get; set; }
+        public Nullable<int> id_session { get; set; }
         public string correo { get; set; }
     
         public virtual user_session user_session { get; set; }
