+++ Proyecto/asp/ForoUdec mayo 2018/Persistence/publicacion.cs
@@ -22,15 +22,15 @@ namespace Persistence
             this.puntuacion = new HashSet<puntuacion>();
         }
     
-        public long id { get; set; }
+        public int id { get; set; }
         public string titulo { get; set; }
         public string descripcion { get; set; }
         public string contenido { get; set; }
-        public Nullable<long> visitas { get; set; }
+        public Nullable<int> visitas { get; set; }
         public Nullable<System.DateTime> fecha { get; set; }
         public string fuente { get; set; }
         public Nullable<int> id_estado { get; set; }
-        public Nullable<long> id_usuario { get; set; }
+        public Nullable<int> id_usuario { get; set; }
         public string miniatura { get; set; }
         public string etiquetas { get; set; }
         public Nullable<int> id_categoria { get; set; }
