+++ Proyecto/asp/ForoUdec mayo 2018/Persistence/puntuacion.cs
@@ -14,11 +14,11 @@ namespace Persistence
     
     public partial class puntuacion
     {
-        public long id { get; set; }
+        public int id { get; set; }
         public Nullable<int> puntuacion1 { get; set; }
         public Nullable<System.DateTime> fecha { get; set; }
-        public Nullable<long> id_usuario { get; set; }
-        public Nullable<long> id_publicacion { get; set; }
+        public Nullable<int> id_usuario { get; set; }
+        public Nullable<int> id_publicacion { get; set; }
     
         public virtual publicacion publicacion { get; set; }
         public virtual user_session user_session { get; set; }
